Local storage data is retrieved before setting initial state in global provider
